+++ libs/shared/src/domains/plan-subscriptions/plan-subscriptions.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, Logger, OnModuleInit } from '@nestjs/common';
 import { CreatePlanSubscriptionDto } from './dto/create-plan-subscription.dto';
 import { PlanSubscription } from './entities/plan-subscription.entity';
-import { PlanSubscriptionsEventsRepository } from './repositories/events/plan-subscriptions-events.repository';
+import { PlanSubscriptionsEventsRepository } from './repositories/events/events.repository';
 
 @Injectable()
 export class PlanSubscriptionsService implements OnModuleInit {
