+++ libs/shared/src/domains/plan-subscriptions/repositories/database/database.repository.ts
@@ -2,8 +2,8 @@ import { Logger } from '@nestjs/common';
 import { featureFlags } from '../../../../config';
 import { CreatePlanSubscriptionDto } from '../../dto/create-plan-subscription.dto';
 import { PlanSubscription } from '../../entities/plan-subscription.entity';
-import { InMemoryPlanSubscriptionsDatabaseRepository } from './implementation/in-memory-plan-subscriptions-database.repository';
-import { PrismaPlanSubscriptionsDatabaseRepository } from './implementation/prisma-plan-subscriptions-database.repository';
+import { InMemoryPlanSubscriptionsDatabaseRepository } from './implementation/in-memory.repository';
+import { PrismaPlanSubscriptionsDatabaseRepository } from './implementation/prisma.repository';
 
 // Abstraction
 export abstract class PlanSubscriptionsDatabaseRepository {
