+++ libs/shared/src/domains/plan-subscriptions/repositories/database/implementation/in-memory.repository.ts
@@ -2,7 +2,7 @@
 import { ConflictException, Injectable } from '@nestjs/common';
 import { CreatePlanSubscriptionDto } from '../../../dto/create-plan-subscription.dto';
 import { PlanSubscription } from '../../../entities/plan-subscription.entity';
-import { PlanSubscriptionsDatabaseRepository } from '../plan-subscriptions-database.repository';
+import { PlanSubscriptionsDatabaseRepository } from '../database.repository';
 
 @Injectable()
 export class InMemoryPlanSubscriptionsDatabaseRepository
