+++ libs/shared/src/domains/plan-subscriptions/repositories/events/implementation/kafka.repository.ts
@@ -2,7 +2,7 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { KafkaEventsService } from '../../../../../infra/events/implementations/kafka.service';
 import { PlanSubscription } from '../../../entities/plan-subscription.entity';
-import { PlanSubscriptionsEventsRepository } from '../plan-subscriptions-events.repository';
+import { PlanSubscriptionsEventsRepository } from '../events.repository';
 import { PLAN_SUBSCRIPTIONS_TOPICS } from '../topics';
 
 @Injectable()
