+++ libs/shared/src/domains/users/repositories/database/implementation/in-memory.repository.ts
@@ -2,7 +2,7 @@
 import { ConflictException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from '../../../dto/create-user.dto';
 import { User } from '../../../entities/user.entity';
-import { UsersDatabaseRepository } from '../users-database.repository';
+import { UsersDatabaseRepository } from '../database.repository';
 
 @Injectable()
 export class InMemoryUsersDatabaseRepository
