--- libs/shared/src/domains/users/repositories/events/dto/types.ts
@@ -1,4 +1,0 @@
-export interface IUserCreatedMessagePayload {
-  id: string;
-  email: string;
-}
