+++ libs/shared/src/domains/users/repositories/events/implementation/in-memory.repository.ts
@@ -1,22 +1,23 @@
 // users.repository.ts
 import { Injectable } from '@nestjs/common';
-import { KafkaEventsService } from '../../../../../infra/events/implementations/kafka.service';
+import { EventsService } from '../../../../../infra/events/events.service';
 import { User } from '../../../entities/user.entity';
+import { UsersEventsRepository } from '../events.repository';
 import { USERS_TOPICS } from '../topics';
-import { UsersEventsRepository } from '../users-events.repository';
 
 @Injectable()
-export class KafkaUsersEventsRepository implements UsersEventsRepository {
-  constructor(private kafkaEventsService: KafkaEventsService) {}
+export class InMemoryUsersEventsRepository implements UsersEventsRepository {
+  constructor(private eventsService: EventsService) {}
 
   async publishUserCreated(user: User): Promise<void> {
-    this.kafkaEventsService.publish({
+    this.eventsService.publish({
       topic: USERS_TOPICS['USER_CREATED'],
       messages: [
         {
+          key: user.email,
           value: JSON.stringify({
-            id: user.id,
             email: user.email,
+            id: user.id,
           }),
         },
       ],
