+++ libs/shared/src/domains/users/repositories/events/implementation/kafka.repository.ts
@@ -1,23 +1,22 @@
 // users.repository.ts
 import { Injectable } from '@nestjs/common';
-import { EventsService } from '../../../../../infra/events/events.service';
+import { KafkaEventsService } from '../../../../../infra/events/implementations/kafka.service';
 import { User } from '../../../entities/user.entity';
+import { UsersEventsRepository } from '../events.repository';
 import { USERS_TOPICS } from '../topics';
-import { UsersEventsRepository } from '../users-events.repository';
 
 @Injectable()
-export class InMemoryUsersEventsRepository implements UsersEventsRepository {
-  constructor(private eventsService: EventsService) {}
+export class KafkaUsersEventsRepository implements UsersEventsRepository {
+  constructor(private kafkaEventsService: KafkaEventsService) {}
 
   async publishUserCreated(user: User): Promise<void> {
-    this.eventsService.publish({
+    this.kafkaEventsService.publish({
       topic: USERS_TOPICS['USER_CREATED'],
       messages: [
         {
-          key: user.email,
           value: JSON.stringify({
-            email: user.email,
             id: user.id,
+            email: user.email,
           }),
         },
       ],
